+++ CHANGES.md
@@ -99,6 +99,7 @@ Changes in 10.0
 - Add MySQL database dump in rosariosis_mysql.sql
 - Log "RosarioSIS HACKING ATTEMPT" into Apache error.log in HackingLog.fnc.php
 - Force URL & menu reloading, always use JS to redirect in HackingLog.fnc.php
+- Place currency symbol after amount for some locales in Currency.fnc.php
 
 Changes in 9.3.1
 ----------------